Spring

+++ demo/src/main/java/com/demo/common/service/spring/learn2/aop/bean/DemoService.java
@@ -2,5 +2,5 @@ package com.demo.common.service.spring.learn2.aop.bean;
 
 public interface DemoService {
 
-    public void queryById(String id, String name, String v1, String v2, Integer age, String v3, String v4, Integer v5);
+    public void queryById(String id, String name, String v1, String v2, Integer age, String v3, String v4, Double v5);
 }
