+++ demo/src/main/java/com/demo/common/service/spring/learn2/aop/bean/DemoServiceImpl.java
@@ -9,7 +9,7 @@ public class DemoServiceImpl implements DemoService {
 
     @MyLogs("queryById-method")
     @Override
-    public void queryById(String id, String name, String v1, String v2, Integer age, String v3, String v4, Integer v5) {
+    public void queryById(String id, String name, String v1, String v2, Integer age, String v3, String v4, Double v5) {
         System.out.println("doSomeThing in " + age + " by " + name + "( by " + id + " ) >>> " + this.hashCode());
     }
 }
